requirements.txt
requests>=2.31.0
python-dotenv==1.0.0
pathvalidate>=3.3.1
pyinstaller>=6.0.0
